requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
imapclient>=2.3.1
python-docx>=1.1.0
pdfplumber>=0.11.0
openpyxl>=3.1.2
xlrd>=2.0.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
httpx>=0.27.0
schedule>=1.2.0
